requirements.txt
cachetools==2.1.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
dill==0.2.8.2
enum34==1.1.6
Flask==1.0.2
future==0.16.0
futures==3.2.0
gapic-google-cloud-datastore-v1==0.15.3
google-api-core==1.2.1
google-auth==1.5.0
google-cloud-core==0.28.1
google-cloud-datastore==1.6.0
google-gax==0.15.16
googleapis-common-protos==1.5.3
grpcio==1.13.0
gunicorn==19.7.1
httplib2==0.11.3
idna==2.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
oauth2client==3.0.0
ply==3.8
proto-google-cloud-datastore-v1==0.90.4
protobuf==3.6.0
pyasn1==0.4.3
pyasn1-modules==0.2.2
pytz==2018.5
requests==2.19.1
romkan==0.2.1
rsa==3.4.2
six==1.11.0
urllib3==1.23
Werkzeug==0.14.1
